Added character controls, scene change and camera switching
Created a temporary character to use for movement and abilities testing. Different areas will now be loaded and unloaded after getting the keypad code correct. Working on implementing a camera switch when interacting close enough with the keypad.

Movement is kind of weird right now an may need to swap to cinemachine cameras later on.